requirements.txt
python-chess>=1.999
pytest>=8.0,<9.0

